+++ packages/deploy/scripts/deploy.ts
@@ -1,5 +1,5 @@
 import { ethers } from "ethers";
-import { NocturneDeployer } from "../src/deploy";
+import { deployNocturne } from "../src/deploy";
 import * as dotenv from "dotenv";
 import * as fs from "fs";
 import { checkNocturneContractDeployment } from "../src/checks";
@@ -10,24 +10,42 @@ dotenv.config();
 
 (async () => {
   const proxyAdminOwner = process.env.PROXY_ADMIN_OWNER;
-  if (!proxyAdminOwner) throw new Error("Missing proxy admin owner address");
+  if (!proxyAdminOwner) throw new Error("Missing PROXY_ADMIN_OWNER");
+
+  const walletOwner = process.env.WALLET_OWNER;
+  if (!walletOwner) throw new Error("Missing WALLET_OWNER");
+
+  const depositManagerOwner = process.env.DEPOSIT_MANAGER_OWNER;
+  if (!depositManagerOwner) throw new Error("Missing DEPOSIT_MANAGER_OWNER");
+
+  const screenersString = process.env.SCREENERS;
+  if (!screenersString) throw new Error("Missing SCREENERS");
+  const screeners = screenersString?.split(",") ?? [];
 
   const useMockSubtreeUpdateVerifier =
     process.env.USE_MOCK_SUBTREE_UPDATE_VERIFIER != undefined;
 
   const deployerKey = process.env.DEPLOYER_KEY;
-  if (!deployerKey) throw new Error("Missing deployer key");
+  if (!deployerKey) throw new Error("Missing DEPLOYER_KEY");
 
   const rpcUrl = process.env.RPC_URL;
-  if (!rpcUrl) throw new Error("Missing rpc url");
+  if (!rpcUrl) throw new Error("Missing RPC_URL");
 
   const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
   const deployer = new ethers.Wallet(deployerKey, provider);
 
-  const nocturneDeployer = new NocturneDeployer(deployer);
-  const deployment = await nocturneDeployer.deployNocturne(proxyAdminOwner, {
-    useMockSubtreeUpdateVerifier,
-  });
+  const deployment = await deployNocturne(
+    deployer,
+    {
+      proxyAdminOwner,
+      walletOwner,
+      depositManagerOwner,
+      screeners,
+    },
+    {
+      useMockSubtreeUpdateVerifier,
+    }
+  );
 
   console.log(deployment);