+++ packages/deploy/src/deploy.ts
@@ -9,6 +9,7 @@ import {
   TransparentUpgradeableProxy__factory,
   Vault__factory,
   Wallet__factory,
+  DepositManager__factory,
 } from "@nocturne-xyz/contracts";
 import { ethers } from "ethers";
 import { ProxiedContract } from "./proxy";
@@ -19,6 +20,13 @@ import {
 } from "@nocturne-xyz/config";
 import { Address } from "./utils";
 
+export interface NocturneDeployArgs {
+  proxyAdminOwner: Address;
+  walletOwner: Address;
+  depositManagerOwner: Address;
+  screeners: Address[];
+}
+
 export interface NocturneDeployOpts {
   proxyAdmin?: ProxyAdmin;
   provider?: any; // FIX: ts build within hh disallows ethers.providers.Provider
@@ -26,128 +34,167 @@ export interface NocturneDeployOpts {
   confirmations?: number;
 }
 
-export class NocturneDeployer {
-  connectedSigner: ethers.Wallet;
+export async function deployNocturne(
+  connectedSigner: ethers.Wallet,
+  args: NocturneDeployArgs,
+  opts?: NocturneDeployOpts
+): Promise<NocturneContractDeployment> {
+  if (!connectedSigner.provider)
+    throw new Error("Wallet must be connected to provider");
 
-  constructor(connectedSigner: ethers.Wallet) {
-    if (!connectedSigner.provider)
-      throw new Error("Wallet must be connected to provider");
-    this.connectedSigner = connectedSigner;
+  const { name, chainId } = await connectedSigner.provider.getNetwork();
+  const startBlock = await connectedSigner.provider.getBlockNumber();
+
+  // Maybe deploy proxy admin
+  let proxyAdmin = opts?.proxyAdmin;
+  if (!proxyAdmin) {
+    console.log("\nDeploying ProxyAdmin");
+    proxyAdmin = await new ProxyAdmin__factory(connectedSigner).deploy();
+    const tx = await proxyAdmin.transferOwnership(args.proxyAdminOwner);
+    await tx.wait(opts?.confirmations);
   }
 
-  async deployNocturne(
-    proxyAdminOwner: Address,
-    opts?: NocturneDeployOpts
-  ): Promise<NocturneContractDeployment> {
-    const { name, chainId } = await this.connectedSigner.provider.getNetwork();
-    const startBlock = await this.connectedSigner.provider.getBlockNumber();
+  // Deploy vault proxy, un-initialized
+  console.log("\nDeploying proxied Vault");
+  const proxiedVault = await deployProxiedContract(
+    new Vault__factory(connectedSigner),
+    proxyAdmin
+  );
+  console.log("Deployed proxied Vault:", proxiedVault.proxyAddresses);
 
-    // Maybe deploy proxy admin
-    let proxyAdmin = opts?.proxyAdmin;
-    if (!proxyAdmin) {
-      console.log("\nDeploying ProxyAdmin");
-      proxyAdmin = await new ProxyAdmin__factory(this.connectedSigner).deploy();
-      const tx = await proxyAdmin.transferOwnership(proxyAdminOwner);
-      await tx.wait(opts?.confirmations);
-    }
+  console.log("\nDeploying JoinSplitVerifier");
+  const joinSplitVerifier = await new JoinSplitVerifier__factory(
+    connectedSigner
+  ).deploy();
+  await joinSplitVerifier.deployTransaction.wait(opts?.confirmations);
 
-    // Deploy vault proxy, un-initialized
-    console.log("\nDeploying proxied Vault");
-    const proxiedVault = await this.deployProxiedContract(
-      new Vault__factory(this.connectedSigner),
-      proxyAdmin
-    );
-    console.log("Deployed proxied Vault:", proxiedVault.proxyAddresses);
-
-    console.log("\nDeploying JoinSplitVerifier");
-    const joinSplitVerifier = await new JoinSplitVerifier__factory(
-      this.connectedSigner
+  let subtreeUpdateVerifier: SubtreeUpdateVerifier | TestSubtreeUpdateVerifier;
+  console.log("\nDeploying SubtreeUpdateVerifier");
+  if (opts?.useMockSubtreeUpdateVerifier) {
+    subtreeUpdateVerifier = await new TestSubtreeUpdateVerifier__factory(
+      connectedSigner
     ).deploy();
-    await joinSplitVerifier.deployTransaction.wait(opts?.confirmations);
+  } else {
+    subtreeUpdateVerifier = await new SubtreeUpdateVerifier__factory(
+      connectedSigner
+    ).deploy();
+  }
+  await subtreeUpdateVerifier.deployTransaction.wait(opts?.confirmations);
 
-    let subtreeUpdateVerifier:
-      | SubtreeUpdateVerifier
-      | TestSubtreeUpdateVerifier;
-    console.log("\nDeploying SubtreeUpdateVerifier");
-    if (opts?.useMockSubtreeUpdateVerifier) {
-      subtreeUpdateVerifier = await new TestSubtreeUpdateVerifier__factory(
-        this.connectedSigner
-      ).deploy();
-    } else {
-      subtreeUpdateVerifier = await new SubtreeUpdateVerifier__factory(
-        this.connectedSigner
-      ).deploy();
-    }
-    await subtreeUpdateVerifier.deployTransaction.wait(opts?.confirmations);
+  console.log("\nDeploying proxied Wallet");
+  const proxiedWallet = await deployProxiedContract(
+    new Wallet__factory(connectedSigner),
+    proxyAdmin,
+    [
+      proxiedVault.address,
+      joinSplitVerifier.address,
+      subtreeUpdateVerifier.address,
+    ]
+  );
+  console.log("Deployed proxied Wallet:", proxiedWallet.proxyAddresses);
 
-    console.log("\nDeploying proxied Wallet");
-    const proxiedWallet = await this.deployProxiedContract(
-      new Wallet__factory(this.connectedSigner),
-      proxyAdmin,
-      [
-        proxiedVault.address,
-        joinSplitVerifier.address,
-        subtreeUpdateVerifier.address,
-      ]
+  console.log("Initializing proxied Vault");
+  const vaultInitTx = await proxiedVault.contract.initialize(
+    proxiedWallet.address
+  );
+  await vaultInitTx.wait(opts?.confirmations);
+
+  console.log("\nDeploying proxied DepositManager");
+  const proxiedDepositManager = await deployProxiedContract(
+    new DepositManager__factory(connectedSigner),
+    proxyAdmin,
+    [
+      "NocturneDepositManager",
+      "v1",
+      proxiedWallet.address,
+      proxiedVault.address,
+    ]
+  );
+
+  console.log("Setting deposit manager screeners\n");
+  for (const screener of args.screeners) {
+    const tx = await proxiedDepositManager.contract.setScreenerPermission(
+      screener,
+      true
     );
-    console.log("Deployed proxied Wallet:", proxiedWallet.proxyAddresses);
-
-    console.log("Initializing proxied Vault");
-    const vaultInitTx = await proxiedVault.contract.initialize(
-      proxiedWallet.address
-    );
-    await vaultInitTx.wait(opts?.confirmations);
-
-    return {
-      network: {
-        name,
-        chainId,
-      },
-      startBlock,
-      proxyAdminOwner,
-      proxyAdmin: proxyAdmin.address,
-      walletProxy: proxiedWallet.proxyAddresses,
-      vaultProxy: proxiedVault.proxyAddresses,
-      joinSplitVerifier: joinSplitVerifier.address,
-      subtreeUpdateVerifier: subtreeUpdateVerifier.address,
-    };
+    await tx.wait(opts?.confirmations);
   }
 
-  async deployProxiedContract<
-    F extends ethers.ContractFactory,
-    C extends Awaited<ReturnType<F["deploy"]>>
-  >(
-    implementationFactory: F,
-    proxyAdmin: ProxyAdmin,
-    initArgs?: Parameters<C["initialize"]>,
-    opts?: NocturneDeployOpts
-  ): Promise<ProxiedContract<C, TransparentProxyAddresses>> {
-    const implementation = await implementationFactory.deploy();
+  console.log("Adding deposit manager to wallet deposit sources\n");
+  const enrollDepositManagerTx =
+    await proxiedWallet.contract.setDepositSourcePermission(
+      proxiedDepositManager.address,
+      true
+    );
+  await enrollDepositManagerTx.wait(opts?.confirmations);
 
-    // If no init args provided, then set init data to empty
-    let initData = "0x";
-    if (initArgs) {
-      initData = implementation.interface.encodeFunctionData(
-        "initialize",
-        initArgs
-      );
-    }
+  console.log("Relinquishing control of wallet and deposit manager\n");
+  const walletTransferOwnershipTx =
+    await proxiedWallet.contract.transferOwnership(args.walletOwner);
+  await walletTransferOwnershipTx.wait(opts?.confirmations);
 
-    const proxyConstructorArgs: Parameters<
-      TransparentUpgradeableProxy__factory["deploy"]
-    > = [implementation.address, proxyAdmin.address, initData];
-    const proxy = await new TransparentUpgradeableProxy__factory(
-      this.connectedSigner
-    ).deploy(...proxyConstructorArgs);
-    await proxy.deployTransaction.wait(opts?.confirmations);
+  const depositManagerTransferOwnershipTx =
+    await proxiedDepositManager.contract.transferOwnership(
+      args.depositManagerOwner
+    );
+  await depositManagerTransferOwnershipTx.wait(opts?.confirmations);
 
-    return new ProxiedContract<C, TransparentProxyAddresses>(
-      implementation.attach(proxy.address) as C,
-      {
-        kind: ProxyKind.Transparent,
-        proxy: proxy.address,
-        implementation: implementation.address,
-      }
+  return {
+    network: {
+      name,
+      chainId,
+    },
+    startBlock,
+    owners: {
+      proxyAdminOwner: args.proxyAdminOwner,
+      walletOwner: args.walletOwner,
+      depositManagerOwner: args.depositManagerOwner,
+    },
+    proxyAdmin: proxyAdmin.address,
+    depositManagerProxy: proxiedDepositManager.proxyAddresses,
+    walletProxy: proxiedWallet.proxyAddresses,
+    vaultProxy: proxiedVault.proxyAddresses,
+    joinSplitVerifierAddress: joinSplitVerifier.address,
+    subtreeUpdateVerifierAddress: subtreeUpdateVerifier.address,
+    screeners: args.screeners,
+    depositSources: [proxiedDepositManager.address],
+  };
+}
+
+async function deployProxiedContract<
+  F extends ethers.ContractFactory,
+  C extends Awaited<ReturnType<F["deploy"]>>
+>(
+  implementationFactory: F,
+  proxyAdmin: ProxyAdmin,
+  initArgs?: Parameters<C["initialize"]>,
+  opts?: NocturneDeployOpts
+): Promise<ProxiedContract<C, TransparentProxyAddresses>> {
+  const implementation = await implementationFactory.deploy();
+
+  // If no init args provided, then set init data to empty
+  let initData = "0x";
+  if (initArgs) {
+    initData = implementation.interface.encodeFunctionData(
+      "initialize",
+      initArgs
     );
   }
+
+  const proxyConstructorArgs: Parameters<
+    TransparentUpgradeableProxy__factory["deploy"]
+  > = [implementation.address, proxyAdmin.address, initData];
+  const proxy = await new TransparentUpgradeableProxy__factory(
+    implementationFactory.signer
+  ).deploy(...proxyConstructorArgs);
+  await proxy.deployTransaction.wait(opts?.confirmations);
+
+  return new ProxiedContract<C, TransparentProxyAddresses>(
+    implementation.attach(proxy.address) as C,
+    {
+      kind: ProxyKind.Transparent,
+      proxy: proxy.address,
+      implementation: implementation.address,
+    }
+  );
 }