+++ packages/e2e-tests/tests/EndToEnd.ts
@@ -3,11 +3,12 @@ import { setupTestDeployment, setupTestClient } from "../src/deploy";
 import { KEYS_TO_WALLETS } from "../src/keys";
 import { ethers } from "ethers";
 import {
+  DepositManager,
   SimpleERC1155Token__factory,
   SimpleERC20Token__factory,
   SimpleERC721Token__factory,
-  Vault,
   Wallet,
+  Vault,
 } from "@nocturne-xyz/contracts";
 import { SimpleERC20Token } from "@nocturne-xyz/contracts/dist/src/SimpleERC20Token";
 import { SimpleERC721Token } from "@nocturne-xyz/contracts/dist/src/SimpleERC721Token";
@@ -46,8 +47,9 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
   let aliceEoa: ethers.Wallet;
   let bobEoa: ethers.Wallet;
 
-  let vault: Vault;
+  let depositManager: DepositManager;
   let wallet: Wallet;
+  let vault: Vault;
   let erc20Token: SimpleERC20Token;
   let erc721Token: SimpleERC721Token;
   let erc1155Token: SimpleERC1155Token;
@@ -65,7 +67,7 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
       },
     });
 
-    ({ provider, teardown, wallet, vault } = testDeployment);
+    ({ provider, teardown, wallet, vault, depositManager } = testDeployment);
 
     const [deployerEoa, _aliceEoa, _bobEoa] = KEYS_TO_WALLETS(provider);
 
@@ -125,8 +127,7 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
   it(`Alice deposits two ${PER_NOTE_AMOUNT} token notes, unwraps ${ALICE_UNWRAP_VAL} tokens publicly, ERC20 transfers ${ALICE_TO_BOB_PUB_VAL} to Bob, and pays ${ALICE_TO_BOB_PRIV_VAL} to Bob privately`, async () => {
     console.log("Deposit funds and commit note commitments");
     await depositFunds(
-      wallet,
-      vault,
+      depositManager,
       erc20Token,
       aliceEoa,
       nocturneWalletSDKAlice.signer.generateRandomStealthAddress(),
@@ -224,8 +225,7 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
   it(`Alice mints an ERC721 and ERC1155 and receives them privately them as refunds to her Nocturne address`, async () => {
     console.log("Deposit funds and commit note commitments");
     await depositFunds(
-      wallet,
-      vault,
+      depositManager,
       erc20Token,
       aliceEoa,
       nocturneWalletSDKAlice.signer.canonicalStealthAddress(),