+++ packages/config/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ### Unreleased
 
+- Add checks in single unit test to ensure fields are defined
+- Add `depositManager`, `depositSources`, `screeners`, and `owners` (added `walletOwner` and `depositManagerOwner` fields)
 - Start `@nocturne-xyz/config` package